requirements.txt
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
influxdb-client==1.49.0
iniconfig==2.3.0
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reactivex==4.1.0
requests==2.32.5
selenium==4.39.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
